requirements.txt
discord.py
python-dotenv
certifi
beautifulsoup4
aiohttp
